fix epsilon to float

+++ flow_main.py
@@ -36,7 +36,7 @@ def parse_arguments():
     parser.add_argument("-p", "--ploidy", type=int, default=2, help="Ploidy of the dataset. (default:%(default)s)")
     parser.add_argument("-S", "--super_prob", type=int, default=-10, help="Probability for using the super edges when there are other edges available. (default:%(default)s)")
     parser.add_argument("-s", "--cheap_prob", type=int, default=-2, help="Probability for using the super edges when there's no other edge available. (default:%(default)s)")
-    parser.add_argument("-e", "--epsilon", type=int, default=0.3, help="Epsilon value for adjusting CN0 counts to probabilities (default:%(default)s)")
+    parser.add_argument("-e", "--epsilon", type=float, default=0.3, help="Epsilon value for adjusting CN0 counts to probabilities (default:%(default)s)")
     parser.add_argument("-b", "--bin_size", default=100, type=int, help="Set the bin size to use for the NB parameters' estimation. (default:%(default)s)")
     parser.add_argument("-d", "--pickle", type=str, help="Pickle dump with the data.", required=False)
